+++ ui/js/component/modalAuthFailure/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { connect } from "react-redux";
+import { doCloseModal } from "actions/app";
+import ModalAuthFailure from "./view";
+
+const select = state => ({});
+
+const perform = dispatch => ({
+  close: () => dispatch(doCloseModal()),
+});
+
+export default connect(select, perform)(ModalAuthFailure);